CloudFunction/requirements.txt
# Function dependencies, for example:
# package>=version
mecab-python3>=0.7
beautifulsoup4
